Bring attention type into variations
This separates the attention type into the variations/ folder so we can
experiment with different types of attention.

Next steps are adding mixtures of attention like GPT3 sliding window +
dense attention, or NVidia's Hymba, then possibly new variations
altogether.
